requirements.txt
Click==7.0
coverage==4.5.3
coveralls==1.6.0
pytest==4.3.1
pytest-cov==2.6.1
Django==2.1.7
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django-debug-toolbar==1.11
django-extensions==2.1.6
django-heroku==0.3.1
django-storages==1.7.1
django-rest-swagger==2.2.0
djangorestframework==3.9.2
gunicorn==19.9.0
pip-tools==3.5.0
psycopg2==2.7.7
PyJWT==1.7.
six==1.12.0
whitenoise==4.1.2
dj-database-url==0.5.0
Pillow==5.4.1
sendgrid==3.6.5
sendgrid-django==4.2.0
django-cloudinary-storage==0.2.3
django-filter==2.1.0
django-social-auth==0.7.28
facebook-sdk==3.1.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
oauth2==1.9.0.post1
oauthlib==3.0.1
python-openid==2.2.5
python-twitter==3.5
requests-oauthlib==1.2.0